Написал свой модуль utils и перенести в него функцию currency_rates() из предыдущего задания. скрипт также работает из консоли, запускать task_4_5.py
